Retool projects core
Split projects object and concept out of the
manager module so they are created/updated each
use
