+++ manager/manager_cmds/projects.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2022, National Technology & Engineering Solutions of Sandia,
+# LLC (NTESS). Under the terms of Contract DE-NA0003525 with NTESS, the U.S.
+# Government retains certain rights in this software.
+#
+# This software is released under the BSD 3-clause license. See LICENSE file
+
+import spack.extensions.manager as manager
+
+
+def add(path):
+    current_projects = manager.config_yaml["projects"]
+    if path not in current_projects:
+        current_projects.append(path)
+        manager.write_config()
